microservices/mr_sandwich/src/food_factory/app/requirements.txt
asgiref==3.5.2
click==8.1.3
djagger==1.1.4
Django==4.1.2
django-storages==1.13.1
djangorestframework==3.14.0
djantic==0.7.0
dsnparse==0.1.15
Flask==2.2.2
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pydantic==1.10.2
pytz==2022.5
sqlparse==0.4.3
typing_extensions==4.4.0
Werkzeug==2.2.2
zipp==3.8.1
